Stage 6 confirmed on hardware; narrow target_temperature to int, add device delete endpoint
Stage 6 (telemetry and commands) fully confirmed end-to-end on real
hardware: registration, heartbeat, telemetry, command execution (AC
reacted correctly), invalid-command rejection, and dedup all verified
after a clean DB reset.

Backend changes (explicitly approved):
- target_temperature narrowed from float to int in ClimateState --
  no AC hardware supports fractional-degree setpoints. Updated the
  backend schema, the simulator's mirrored model and internal
  Simulator.target_temperature field, and the firmware
  (BackendCommand::targetTemperature, CommandProcessor's temperature
  bounds, and removal of the now-unneeded float-rounding trick).
- Added DELETE /api/v1/management/devices/{id}, closing the gap
  tracked in TODO.md. Cascades via existing FK ondelete="CASCADE".
- Four further backend suggestions (test coverage, CommandCreate.type
  as Literal, typed CommandResultIn.result, devices-list pagination)
  logged in TODO.md rather than implemented, per instruction.

Firmware: removed the printSerialHelp() boot-time dump now that
backend-driven commands are the primary path -- discoverability is
preserved via the existing unrecognized-key fallback.

+++ backend/app/schemas/device_state.py
@@ -7,7 +7,7 @@ class ClimateState(BaseModel):
     model_config = ConfigDict(from_attributes=True)
 
     power: bool | None = None
-    target_temperature: float | None = None
+    target_temperature: int | None = None
 
 
 class DeviceStateOut(BaseModel):
